requirements.txt
flask
imagehash
llama-cpp-python
matplotlib
numpy==1.26.4
onnxruntime==1.17.3
pillow==10.3.0
redis==5.0.4
